243: move SendRawTransactionResult
